Split off `participant` module
